+++ sic-storage-web/src/main/java/ru/insoft/archive/sic_storage/model/view/VStrgPlaceOrg.java
@@ -42,7 +42,7 @@ public class VStrgPlaceOrg implements JsonOut
     private String phone;
 
     @Column(name = "DOCUMENT_COUNT")
-    private Integer documentCount;
+    private Long documentCount;
 
     @Column(name = "BEGIN_YEAR")
     private Integer beginYear;
@@ -113,11 +113,11 @@ public class VStrgPlaceOrg implements JsonOut
         this.phone = phone;
     }
 
-    public Integer getDocumentCount() {
+    public Long getDocumentCount() {
         return documentCount;
     }
 
-    public void setDocumentCount(Integer documentCount) {
+    public void setDocumentCount(Long documentCount) {
         this.documentCount = documentCount;
     }
